requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
googlemaps==4.10.0
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
validate-email==1.3
beautifulsoup4==4.12.2
flask==3.0.2
gunicorn==21.2.0
Flask-Login==0.6.3
Authlib==1.3.0 